requirements.txt
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==2.3.2
flask-redis==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
python-dotenv==1.0.0
redis==5.0.3
requests==2.31.0
urllib3==2.2.3
Werkzeug==3.0.4
